requirements.txt
absl-py==1.1.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
idna==3.3
importlib-metadata==4.11.4
joblib==1.1.0
Markdown==3.3.7
numpy==1.21.6
oauthlib==3.2.0
Pillow==9.1.1
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytorch-metric-learning==1.3.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
timm==0.5.4
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
